+++ Changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### v1.3.0
+- Allow integration with shops that have Magento_InventoryApi module disabled
+
+  Product stock will be retrieved as 0
+
+- Fix issue with order total
+
 ### v1.2.1
 - Fix possible issue with order total and vouchers
 - Fix issue with wrong shipping price_with_tax value